WCOSS fixes and no-DA option (NODA=YES).

+++ jobs/rocoto/prep.sh
@@ -39,9 +39,10 @@ if [ $ROTDIR_DUMP = "YES" ]; then
     gPDY=$(echo $GDATE | cut -c1-8)
     gcyc=$(echo $GDATE | cut -c9-10)
     GDUMP=wdas
+    PDUMP=gdas
     gCOMOUT="$ROTDIR/$GDUMP.$gPDY/$gcyc"
     if [ ! -s $gCOMOUT/$GDUMP.t${gcyc}z.updated.status.tm00.bufr_d ]; then
-     $HOMEwfs/ush/getdump.sh $GDATE gdas $DMPDIR/${GDUMP}${DUMP_SUFFIX}.${gPDY}/${gcyc} $gCOMOUT
+     $HOMEwfs/ush/getdump.sh $GDATE gdas $DMPDIR/${PDUMP}${DUMP_SUFFIX}.${gPDY}/${gcyc} $gCOMOUT
      status=$?
      [[ $status -ne 0 ]] && exit $status
     fi
@@ -81,6 +82,7 @@ fi
 
 ###############################################################
 # Generate prepbufr files from dumps or copy from OPS
+[ $NODA = "YES" ] && exit
 if [ $DO_MAKEPREPBUFR = "YES" ]; then
     if [ $ROTDIR_DUMP = "YES" ]; then
 	rm $COMOUT/${OPREFIX}prepbufr
